requirements.txt
albumentations==1.3.0
einops==0.6.0
matplotlib
numpy
pandas
progress==1.6
PyYAML==6.0
scikit_learn
torch==1.13.0
tqdm==4.64.1
transforms==0.1
vit_pytorch==0.40.2
tensorboard